Moved client code to it's own Repo
