+++ server/lms-server/lms-server/Repository/IBooksRepository.cs
@@ -14,12 +14,5 @@ namespace lms_server.Repository
 
         Task<Book?> DeleteBookAsync(Guid id);
 
-        Task<List<Client>> GellAllClientsAsync();
-
-        Task<Client> GetClientByID(Guid id);
-
-        Task<Client> UpdateClientDetails(Guid id,Client client);
-
-        Task<Client> DeleteClientAsync(Guid id);
     }
 }
